Calling "Save revision to" in the list of revisions of the log dialog resulted in the proposed file name "filename-1(null)". Correct is "filename-5.txt". Two errors: * revision number was always 1 * extension was replaced by "(null)".
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@8184 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/TortoiseProc/LogDlg.cpp
@@ -3082,7 +3082,7 @@ void CLogDlg::ShowContextMenuForRevisions(CWnd* /*pWnd*/, CPoint point)
 					CString strWinPath = m_path.GetWinPathString();
 					int rfind = strWinPath.ReverseFind('.');
 					if (rfind > 0)
-						revFilename.Format(_T("%s-%ld%s"), (LPCTSTR)strWinPath.Left(rfind), revSelected, (LPCTSTR)strWinPath.Mid(rfind));
+						revFilename.Format(_T("%s-%ld%s"), (LPCTSTR)strWinPath.Left(rfind), (LONG)revSelected, (LPCTSTR)strWinPath.Mid(rfind));
 					else
 						revFilename.Format(_T("%s-%ld"), (LPCTSTR)strWinPath, revSelected);
 					_tcscpy_s(szFile, MAX_PATH, revFilename);
